Move base image to Debian Stretch
Let's go ahead and move up to Debian Stretch as the base image for
building Maru for maru-0.6 / Android O and up. The main reason for doing
this *now* is that blueprints has transitioned from jessie to stretch by
default for maru-0.6, and stretch just won't build on Ubuntu 14.04
because the debootstrap pkg is too old. The only way around it would be
to use two different docker images for the build and I'd rather KISS!

This is a safe move because even Google devs are switching to using
Debian Testing internally:
https://groups.google.com/d/msg/android-building/Tq7cmJTBN9E/I51qOEfdBgAJ

Signed-off-by: Preetam D'Souza <preetamjdsouza@gmail.com>